code/requirements.txt
allure-pytest==2.12.0
Faker==16.8.1
pydantic==1.10.4
pytest==7.2.1
requests==2.28.2
